feat(decision-analyzer)!: split DecisionAnalyzer into namespace facade
Apply the AGENTS.md "Namespace facade for large analyzer classes"
pattern to DecisionAnalyzer (previously 48 methods on a single class).
The top-level class is now a thin orchestrator with data-loading
classmethods, properties and lifecycle helpers; analysis methods live
on two sub-namespace classes attached as instance attributes, matching
the ADMDatamart reference (dm.plot, dm.aggregates, dm.agb, ...).

Each sub-namespace lives in its own module, takes the parent
DecisionAnalyzer as a single `da` argument, and uses TYPE_CHECKING to
avoid circular imports.

BREAKING CHANGES -- 28 public methods (plus 3 private helpers) moved.
See docs/migration-v4-to-v5.md "Namespace reorganisation" for the
complete before/after table.

Methods moved to `da.aggregates` (16):
  - aggregate_remaining_per_stage
  - get_distribution_data
  - get_funnel_data
  - get_decisions_without_actions_data
  - get_funnel_summary
  - get_optionality_data
  - get_optionality_funnel
  - get_action_variation_data
  - get_offer_variability_stats
  - get_offer_quality
  - filtered_action_counts
  - get_trend_data
  - get_filter_component_data
  - get_component_action_impact
  - get_component_drilldown
  - get_ab_test_results

Methods moved to `da.scoring` (12 public + 3 private):
  - re_rank
  - get_selected_group_rank_boundaries
  - get_sensitivity
  - get_thresholding_data
  - priority_component_distribution
  - all_components_distribution
  - get_win_loss_distribution_data
  - get_winning_or_losing_interactions
  - get_win_loss_counts
  - get_win_loss_distributions
  - get_win_distribution_data
  - find_lever_value
  - _remaining_at_stage (private)
  - _winning_from (private)
  - _losing_to (private)

Unchanged on DecisionAnalyzer:
  - from_explainability_extract, from_decision_analyzer (classmethods)
  - All cached_properties / properties (decision_data, sample,
    preaggregated_*_view, arbitration_stage, mandatory_actions,
    color_mappings, stages_from_arbitration_down,
    stages_with_propensity, propensity_validation_warning,
    overview_stats, available_levels, num_sample_interactions,
    stage_to_group_mapping)
  - Lifecycle / filter helpers (set_level, filtered,
    get_available_fields_for_filtering, get_possible_scope_values,
    get_possible_stage_values)
  - `da.plot` (pre-existing namespace, unchanged)

Internal callers updated (223 replacements across 12 files):
  - python/pdstools/decision_analyzer/plots.py (13)
  - python/pdstools/decision_analyzer/DecisionAnalyzer.py (2 cached
    properties that still call into aggregates)
  - 7 Streamlit pages under python/pdstools/app/decision_analyzer/
  - python/tests/test_DecisionAnalyzer.py, test_da_plots.py
  - Two example notebooks

All 1690 tests pass.
