+++ python/pdstools/app/decision_analyzer/pages/7_Optionality_Analysis.py
@@ -86,7 +86,7 @@ with st.container(border=True):
     )
 
     optionality_data_with_trend_per_stage = (
-        st.session_state.decision_data.get_optionality_data(df=filtered_data, by_day=True)
+        st.session_state.decision_data.aggregates.get_optionality_data(df=filtered_data, by_day=True)
         .group_by(["day", st.session_state.decision_data.level])
         .agg(avg_actions=(pl.col("nOffers") * pl.col("Interactions")).sum() / pl.col("Interactions").sum())
         .sort("day")
