+++ python/pdstools/decision_analyzer/DecisionAnalyzer.py
@@ -1,7 +1,5 @@
 # python/pdstools/decision_analyzer/DecisionAnalyzer.py
-from bisect import bisect_left
 from functools import cached_property
-from typing import Literal
 import logging
 import os
 import warnings
@@ -10,6 +8,8 @@ import polars as pl
 import polars.selectors as cs
 
 from .data_read_utils import validate_columns
+from ._aggregates import Aggregates
+from ._scoring import Scoring
 from .plots import Plot
 from .stage_grouping import DISPLAY_NAME_LOOKUP
 from .column_schema import (
@@ -21,7 +21,6 @@ from .utils import (
     apply_filter,
     determine_extract_type,
     get_table_definition,
-    gini_coefficient,
     rename_and_cast_types,
     resolve_aliases,
 )
@@ -64,6 +63,12 @@ class DecisionAnalyzer:
         Either ``"explainability_extract"`` or ``"decision_analyzer"``.
     plot : Plot
         Plot accessor for visualization methods.
+    aggregates : Aggregates
+        Accessor for aggregation queries (funnel, distribution,
+        optionality, action variation, …).
+    scoring : Scoring
+        Accessor for re-ranking, sensitivity, win/loss and lever
+        analysis.
 
     Examples
     --------
@@ -71,6 +76,8 @@ class DecisionAnalyzer:
     >>> da = DecisionAnalyzer.from_explainability_extract("data/sample_explainability_extract.parquet")
     >>> da.overview_stats
     >>> da.plot.sensitivity()
+    >>> da.aggregates.get_funnel_data(scope="Action")
+    >>> da.scoring.get_sensitivity()
     """
 
     # Lazily-set attribute populated when ``sample`` is first accessed.
@@ -250,6 +257,8 @@ class DecisionAnalyzer:
         >>> decision_analyzer = DecisionAnalyzer(raw_data, mandatory_expr=mandatory)
         """
         self.plot = Plot(self)
+        self.aggregates = Aggregates(self)
+        self.scoring = Scoring(self)
         self.level = level
         self.sample_size = sample_size
         self._num_samples = num_samples
@@ -748,7 +757,7 @@ class DecisionAnalyzer:
         the stages remaining.
         """
         self.preaggregated_decision_data_remainingview = (
-            self.aggregate_remaining_per_stage(
+            self.aggregates.aggregate_remaining_per_stage(
                 self.preaggregated_filter_view,
                 list(self.preaggregation_columns),
                 [
@@ -991,1345 +1000,7 @@ class DecisionAnalyzer:
             )
         )
 
-    def get_distribution_data(
-        self,
-        stage: str,
-        grouping_levels: str | list[str],
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Distribution of decisions by grouping columns at a given stage.
-
-        Parameters
-        ----------
-        stage : str
-            Stage to filter on.
-        grouping_levels : str or list of str
-            Column(s) to group by (e.g. ``"Action"`` or ``["Channel", "Action"]``).
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters applied before aggregation.
-
-        Returns
-        -------
-        pl.LazyFrame
-            Columns from *grouping_levels* plus ``Decisions``, sorted descending.
-        """
-        distribution_data = (
-            apply_filter(self.preaggregated_remaining_view, additional_filters)
-            .filter(pl.col(self.level) == stage)
-            .group_by(grouping_levels)
-            .agg(pl.sum("Decisions"))
-            .sort("Decisions", descending=True)
-            .filter(pl.col("Decisions") > 0)
-        )
-
-        return distribution_data
-
-    def get_funnel_data(
-        self, scope: str, additional_filters: pl.Expr | list[pl.Expr] | None = None
-    ) -> tuple[pl.LazyFrame, pl.DataFrame, pl.DataFrame]:
-        filtered_df = apply_filter(self.preaggregated_filter_view, additional_filters)
-
-        interaction_count_expr = (
-            apply_filter(self.decision_data, additional_filters).select("Interaction ID").unique().count()
-        )
-
-        # Available: actions entering each stage (existing logic)
-        available_data = self.aggregate_remaining_per_stage(
-            df=filtered_df,
-            group_by_columns=[scope],
-            aggregations=[
-                pl.sum("Decisions").alias("action_occurrences"),
-                pl.col("Interaction_IDs")
-                .list.explode(keep_nulls=False, empty_as_null=False)
-                .unique()
-                .count()
-                .alias("interaction_count_for_scope"),
-            ],
-        ).filter(pl.col("action_occurrences") > 0)
-
-        # Passing: actions exiting each stage (shift by one — use next stage's remaining set)
-        # Output is the result, not a filtering stage, so exclude it from funnel views
-        stages = [s for s in self.AvailableNBADStages if s != "Output"]
-        # All stages including Output — needed to correctly count actions that survive
-        # past the last filtering stage and reach Output.
-        all_stages = list(self.AvailableNBADStages)
-
-        def _passing_at_stage(i: int, stage: str) -> pl.DataFrame:
-            # For stage i, passing = actions remaining at stage i+1 (including Output).
-            all_stage_idx = all_stages.index(stage)
-            next_stages = all_stages[all_stage_idx + 1 :]
-            if not next_stages:
-                next_stages = [stage]
-            return (
-                filtered_df.filter(pl.col(self.level).is_in(next_stages))
-                .group_by([scope])
-                .agg(
-                    action_occurrences=pl.sum("Decisions"),
-                    interaction_count_for_scope=pl.col("Interaction_IDs")
-                    .list.explode(keep_nulls=False, empty_as_null=False)
-                    .unique()
-                    .count(),
-                )
-                .with_columns(pl.lit(stage).alias(self.level))
-                .collect()
-            )
-
-        passing_data = pl.concat([_passing_at_stage(i, stage) for i, stage in enumerate(stages)]).filter(
-            pl.col("action_occurrences") > 0
-        )
-
-        # Filtered: actions removed at each stage (existing logic, unchanged)
-        filtered_funnel = (
-            filtered_df.filter(pl.col("Record Type") == "FILTERED_OUT")
-            .group_by([self.level, scope])
-            .agg(
-                action_occurrences=pl.sum("Decisions"),
-                interaction_count_for_scope=pl.col("Interaction_IDs")
-                .list.explode(keep_nulls=False, empty_as_null=False)
-                .unique()
-                .count(),
-            )
-            .collect()
-        )
-
-        interaction_count = interaction_count_expr.collect().item()
-        metrics_expr = (
-            pl.lit(interaction_count).alias("interaction_count"),
-            (pl.col("action_occurrences") / pl.lit(interaction_count)).alias("actions_per_interaction"),
-            ((pl.col("interaction_count_for_scope") / pl.lit(interaction_count)) * 100).alias("penetration_pct"),
-        )
-
-        return (
-            available_data.with_columns(metrics_expr),
-            passing_data.with_columns(metrics_expr),
-            filtered_funnel.with_columns(metrics_expr),
-        )
-
-    def get_decisions_without_actions_data(
-        self, additional_filters: pl.Expr | list[pl.Expr] | None = None
-    ) -> pl.DataFrame:
-        """Per-stage count of interactions newly left with no remaining actions.
-
-        Returns a DataFrame with columns [self.level, "decisions_without_actions"],
-        sorted in pipeline order. For each stage X, the value is the number of
-        interactions that lose their final remaining action at stage X.
-        """
-        filter_view = apply_filter(self.preaggregated_filter_view, additional_filters)
-
-        def count_from(stages):
-            return (
-                filter_view.filter(pl.col(self.level).is_in(stages))
-                .select(pl.col("Interaction_IDs").list.explode(keep_nulls=False, empty_as_null=False).unique().count())
-                .collect()
-                .item()
-            )
-
-        total_interactions = (
-            apply_filter(self.decision_data, additional_filters)
-            .select("Interaction ID")
-            .unique()
-            .count()
-            .collect()
-            .item()
-        )
-
-        # Output is the result, not a filtering stage — exclude from this view
-        stages = [s for s in self.AvailableNBADStages if s != "Output"]
-        has_output = "Output" in self.AvailableNBADStages
-
-        cumulative_without_actions: list[float] = []
-        for i, _stage in enumerate(stages):
-            successor_stages = list(stages[i + 1 :])
-            # Include Output in every successor set so survivor sets remain nested,
-            # even when some interactions skip optional intermediate stages.
-            if has_output:
-                successor_stages.append("Output")
-            survivors_after_stage = count_from(successor_stages) if successor_stages else 0
-            cumulative_without_actions.append(float(total_interactions - survivors_after_stage))
-
-        rows = []
-        for i, stage in enumerate(stages):
-            prev_cumulative = cumulative_without_actions[i - 1] if i > 0 else 0.0
-            rows.append(
-                {
-                    self.level: stage,
-                    "decisions_without_actions": cumulative_without_actions[i] - prev_cumulative,
-                }
-            )
-
-        return pl.DataFrame(rows)
-
-    def get_funnel_summary(
-        self,
-        available_df: pl.LazyFrame,
-        passing_df: pl.DataFrame,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.DataFrame:
-        """Per-stage summary: Available, Passing, Filtered actions and Decisions.
-
-        The table matches the funnel chart: it starts with a synthetic
-        "Available Actions" row and excludes the Output stage.
-
-        Parameters
-        ----------
-        available_df : pl.LazyFrame
-            First element returned by ``get_funnel_data`` (actions entering each stage).
-        passing_df : pl.DataFrame
-            Second element returned by ``get_funnel_data`` (actions exiting each stage).
-        additional_filters : optional
-            Same filters used when calling ``get_funnel_data``.
-
-        Returns
-        -------
-        pl.DataFrame
-            One row per stage in pipeline order with raw counts first,
-            then per-decision averages.
-        """
-        level = self.level
-        stages = [s for s in self.AvailableNBADStages if s != "Output"]
-        first_real_stage = stages[0] if stages else None
-
-        available_by_stage = (
-            available_df.collect()
-            .with_columns(pl.col(level).cast(pl.Utf8))
-            .group_by(level)
-            .agg(pl.sum("action_occurrences").alias("Available Actions"))
-        )
-        passing_by_stage = (
-            passing_df.with_columns(pl.col(level).cast(pl.Utf8))
-            .group_by(level)
-            .agg(pl.sum("action_occurrences").alias("Passing Actions"))
-        )
-        without_actions = self.get_decisions_without_actions_data(additional_filters)
-        total_interactions = (
-            apply_filter(self.decision_data, additional_filters)
-            .select("Interaction ID")
-            .unique()
-            .count()
-            .collect()
-            .item()
-        )
-        stage_order_no_output = {s: i for i, s in enumerate(stages)}
-        decisions_by_stage = (
-            without_actions.with_columns(pl.col(level).cast(pl.Utf8))
-            .sort(pl.col(level).replace_strict(stage_order_no_output, default=999, return_dtype=pl.Int32))
-            .with_columns(pl.col("decisions_without_actions").cum_sum().alias("_cumulative_without_actions"))
-            .with_columns((pl.lit(total_interactions) - pl.col("_cumulative_without_actions")).alias("Decisions"))
-            .select([level, "Decisions"])
-        )
-
-        summary = (
-            available_by_stage.join(passing_by_stage, on=level, how="left")
-            .join(decisions_by_stage, on=level, how="left")
-            .fill_null(0)
-            .with_columns((pl.col("Available Actions") - pl.col("Passing Actions")).alias("Filtered Actions"))
-        )
-
-        # Exclude Output, keep only funnel stages
-        summary = summary.filter(pl.col(level).is_in(stages))
-
-        # Add synthetic "Available Actions" baseline row — but only when the first
-        # real stage isn't already called "Available Actions" (to avoid a duplicate).
-        synthetic_label = "Available Actions"
-        if first_real_stage is not None and first_real_stage != synthetic_label:
-            first_available = summary.filter(pl.col(level) == first_real_stage).select("Available Actions").item()
-            synthetic_row = pl.DataFrame(
-                {
-                    level: [synthetic_label],
-                    "Available Actions": [first_available],
-                    "Passing Actions": [first_available],
-                    "Decisions": [total_interactions],
-                    "Filtered Actions": [0],
-                }
-            )
-            summary = pl.concat([synthetic_row, summary], how="diagonal_relaxed")
-
-        display_stage_order = [synthetic_label] + stages
-        display_stage_index = {name: idx for idx, name in enumerate(display_stage_order)}
-        summary = (
-            summary.with_columns(
-                (pl.col("Filtered Actions") / pl.col("Filtered Actions").sum() * 100)
-                .round(1)
-                .alias("% of Total Filtered"),
-                (pl.col("Available Actions") / pl.lit(total_interactions)).round(2).alias("Avg Available/Decision"),
-                (pl.col("Passing Actions") / pl.lit(total_interactions)).round(2).alias("Avg Passing/Decision"),
-                (pl.col("Filtered Actions") / pl.lit(total_interactions)).round(2).alias("Avg Filtered/Decision"),
-                ((pl.lit(total_interactions) - pl.col("Decisions")) / pl.lit(total_interactions) * 100)
-                .round(1)
-                .alias("% Decisions without Actions"),
-            )
-            .sort(pl.col(level).replace_strict(display_stage_index, default=999, return_dtype=pl.Int32))
-            .select(
-                level,
-                "Available Actions",
-                "Passing Actions",
-                "Filtered Actions",
-                "Decisions",
-                "% of Total Filtered",
-                "Avg Available/Decision",
-                "Avg Passing/Decision",
-                "Avg Filtered/Decision",
-                "% Decisions without Actions",
-            )
-        )
-        return summary
-
-    def get_filter_component_data(
-        self, top_n: int, additional_filters: pl.Expr | list[pl.Expr] | None = None
-    ) -> pl.DataFrame:
-        """Top-N filter components per stage, ranked by filtered-decision count.
-
-        Parameters
-        ----------
-        top_n : int
-            Maximum number of components to return per stage.
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters applied before aggregation.
-
-        Returns
-        -------
-        pl.DataFrame
-            Columns include the stage level, Component Name, and
-            Filtered Decisions.
-        """
-        group_cols = [self.level, "Component Name"]
-        available = set(self.preaggregated_filter_view.collect_schema().names())
-        if "Component Type" in available:
-            group_cols.append("Component Type")
-
-        stages_actions_df = (
-            apply_filter(self.preaggregated_filter_view, additional_filters)
-            .filter(pl.col(self.level) != "Output")
-            .group_by(group_cols)
-            .agg(
-                pl.col("Interaction_IDs")
-                .list.explode(keep_nulls=False, empty_as_null=False)
-                .unique()
-                .count()
-                .alias("Filtered Decisions")
-            )
-            .collect()
-        )
-        result = pl.concat(
-            pl.collect_all(
-                [
-                    x.lazy().top_k(top_n, by="Filtered Decisions").sort("Filtered Decisions", descending=False)
-                    for x in stages_actions_df.partition_by(self.level)
-                ]
-            )
-        ).with_columns(pl.col(pl.Categorical).cast(pl.Utf8))
-
-        return result
-
-    def get_component_action_impact(
-        self,
-        top_n: int = 10,
-        scope: str = "Action",
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.DataFrame:
-        """Per-component breakdown of which items are filtered and how many.
-
-        For each component, returns the top-N items (at the chosen scope
-        granularity) it filters out. The scope controls whether the breakdown
-        is at Issue, Group, or Action level.
-
-        Parameters
-        ----------
-        top_n : int, default 10
-            Maximum number of items to return per component.
-        scope : str, default "Action"
-            Granularity level: ``"Issue"``, ``"Group"``, or ``"Action"``.
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters to apply before aggregation.
-
-        Returns
-        -------
-        pl.DataFrame
-            Columns include Component Name, StageGroup, scope columns, and
-            Filtered Decisions. Sorted by component then descending count.
-        """
-        filtered_data = apply_filter(self.decision_data, additional_filters).filter(
-            pl.col("Record Type") == "FILTERED_OUT"
-        )
-
-        # Build scope columns up to and including the requested level
-        scope_hierarchy = ["Issue", "Group", "Action"]
-        scope_idx = scope_hierarchy.index(scope) if scope in scope_hierarchy else 2
-        scope_cols = scope_hierarchy[: scope_idx + 1]
-
-        group_cols = ["Component Name", self.level] + scope_cols
-        available = set(filtered_data.collect_schema().names())
-        group_cols = [c for c in group_cols if c in available]
-
-        impact_df = filtered_data.group_by(group_cols).agg(pl.len().alias("Filtered Decisions")).collect()
-
-        # Top-N items per component
-        result = pl.concat(
-            pl.collect_all(
-                [
-                    part.lazy().top_k(top_n, by="Filtered Decisions").sort("Filtered Decisions", descending=True)
-                    for part in impact_df.partition_by("Component Name")
-                ]
-            )
-        ).with_columns(pl.col(pl.Categorical).cast(pl.Utf8))
-
-        return result.sort(["Component Name", "Filtered Decisions"], descending=[False, True])
-
-    def get_component_drilldown(
-        self,
-        component_name: str,
-        scope: str = "Action",
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-        sort_by: str = "Filtered Decisions",
-    ) -> pl.DataFrame:
-        """Deep-dive into a single filter component showing dropped actions and
-        their potential value.
-
-        Since scoring columns (Priority, Value, Propensity) are typically null
-        on FILTERED_OUT rows, this method derives the action's "potential value"
-        by looking up average scores from rows where the same action survives
-        (non-null Priority/Value). This gives the "value of what's being
-        dropped" perspective.
-
-        Parameters
-        ----------
-        component_name : str
-            The Component Name to drill into.
-        scope : str, default "Action"
-            Granularity level: ``"Issue"``, ``"Group"``, or ``"Action"``.
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters to apply before aggregation.
-        sort_by : str, default "Filtered Decisions"
-            Column to sort results by (descending).
-
-        Returns
-        -------
-        pl.DataFrame
-            Columns include scope columns, Filtered Decisions,
-            avg_Priority, avg_Value, avg_Propensity, Component Type (if
-            available).
-        """
-        base = apply_filter(self.decision_data, additional_filters)
-        available = set(base.collect_schema().names())
-
-        # Filtered rows for this component
-        filtered_rows = base.filter(
-            (pl.col("Record Type") == "FILTERED_OUT") & (pl.col("Component Name") == component_name)
-        )
-
-        scope_hierarchy = ["Issue", "Group", "Action"]
-        scope_idx = scope_hierarchy.index(scope) if scope in scope_hierarchy else 2
-        scope_cols = scope_hierarchy[: scope_idx + 1]
-
-        group_cols = [c for c in scope_cols if c in available]
-        agg_exprs = [pl.len().alias("Filtered Decisions")]
-        if "Component Type" in available:
-            group_cols.append("Component Type")
-
-        filtered_agg = filtered_rows.group_by(group_cols).agg(agg_exprs).collect()
-
-        # Reference scores from surviving rows (non-null Priority)
-        score_cols = ["Priority", "Value", "Propensity"]
-        present_scores = [c for c in score_cols if c in available]
-
-        if present_scores:
-            score_aggs = [pl.col(c).mean().alias(f"avg_{c}") for c in present_scores]
-            join_cols = [c for c in scope_cols if c in available]
-            reference_scores = (
-                base.filter(pl.col("Priority").is_not_null()).group_by(join_cols).agg(score_aggs).collect()
-            )
-            result = filtered_agg.join(
-                reference_scores,
-                on=join_cols,
-                how="left",
-            )
-        else:
-            result = filtered_agg
-
-        sort_col = sort_by if sort_by in result.columns else "Filtered Decisions"
-        return result.with_columns(pl.col(pl.Categorical).cast(pl.Utf8)).sort(sort_col, descending=True)
-
-    def re_rank(
-        self,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-        overrides: list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Recalculate priority and rank for all PVCL component combinations.
-
-        Computes five alternative priority scores by selectively dropping one
-        component at a time (Propensity, Value, Context Weight, Levers) and
-        ranks actions within each interaction for each variant.  This is the
-        foundation for sensitivity analysis.
-
-        Parameters
-        ----------
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Filters applied to the sample before ranking.
-        overrides : list of pl.Expr, optional
-            Column override expressions applied before priority calculation
-            (e.g. to simulate lever adjustments).
-
-        Returns
-        -------
-        pl.LazyFrame
-            Sample data augmented with ``prio_*`` and ``rank_*`` columns
-            for each PVCL variant.
-        """
-        rank_exprs = [
-            pl.struct(
-                [
-                    "is_mandatory",
-                    x,
-                    "Stage Order",
-                    pl.col("Issue").rank() * -1,
-                    pl.col("Group").rank() * -1,
-                    pl.col("Action").rank() * -1,
-                ]
-            )
-            .rank(descending=True, method="ordinal", seed=1)
-            .over(["Interaction ID"])
-            .cast(pl.Int16)
-            .alias(f"rank_{x.split('_')[1]}")
-            for x in ["prio_PVCL", "prio_VCL", "prio_PCL", "prio_PVL", "prio_PVC"]
-        ]
-
-        # Fill missing PVCL components with 1.0 (neutral for multiplication).
-        available = set(self.sample.collect_schema().names())
-        pvcl_defaults = {
-            "Propensity": 1.0,
-            "Value": 1.0,
-            "Context Weight": 1.0,
-            "Levers": 1.0,
-        }
-        fill_exprs = []
-        for col_name, default in pvcl_defaults.items():
-            if col_name in available:
-                fill_exprs.append(pl.col(col_name).fill_null(default))
-            else:
-                fill_exprs.append(pl.lit(default).alias(col_name))
-
-        rank_df = (
-            apply_filter(
-                self.sample.with_columns(fill_exprs),
-                additional_filters,
-            )
-            .with_columns(overrides if overrides is not None else [])
-            .filter(pl.col("Priority").is_not_null())
-            .with_columns(
-                prio_PVCL=(pl.col("Propensity") * pl.col("Value") * pl.col("Context Weight") * pl.col("Levers")),
-                prio_VCL=(pl.col("Value") * pl.col("Context Weight") * pl.col("Levers")),
-                prio_PCL=(pl.col("Propensity") * pl.col("Context Weight") * pl.col("Levers")),
-                prio_PVL=(pl.col("Propensity") * pl.col("Value") * pl.col("Levers")),
-                prio_PVC=(pl.col("Propensity") * pl.col("Value") * pl.col("Context Weight")),
-            )
-            .with_columns(*rank_exprs)
-        )
-
-        return rank_df
-
-    def get_selected_group_rank_boundaries(
-        self,
-        group_filter: pl.Expr | list[pl.Expr],
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Compute selected-group rank boundaries per interaction.
-
-        For each interaction where the selected comparison group is present,
-        returns the best (lowest) and worst (highest) rank observed for the
-        selected rows in arbitration-relevant stages.
-        """
-        selected_rows = (
-            apply_filter(apply_filter(self.sample, additional_filters), group_filter)
-            .filter(pl.col(self.level).is_in(self.stages_from_arbitration_down))
-            .select(["Interaction ID", "Rank"])
-        )
-
-        return selected_rows.group_by("Interaction ID").agg(
-            selected_group_best_rank=pl.col("Rank").min(),
-            selected_group_worst_rank=pl.col("Rank").max(),
-            selected_group_row_count=pl.len(),
-        )
-
     # TODO consider making this more generic by returning all rank views in one pass.
-    def get_win_loss_distribution_data(
-        self,
-        level: str | list[str],
-        win_rank: int | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-        group_filter: pl.Expr | list[pl.Expr] | None = None,
-        status: Literal["Wins", "Losses"] | None = None,
-        top_k: int | None = None,
-    ) -> pl.LazyFrame:
-        """Win/loss distribution at a given scope level.
-
-        Operates in two modes depending on whether *group_filter* is provided:
-
-        * **Without group_filter** (rank-based): uses pre-aggregated data and
-          a fixed *win_rank* threshold to split wins/losses.
-        * **With group_filter** (group-based): uses sample data and
-          per-interaction rank boundaries of the selected group to identify
-          actions it beats (*Wins*) or loses to (*Losses*).
-
-        Parameters
-        ----------
-        level : str or list of str
-            Column(s) to group the distribution by (e.g. ``"Action"``).
-        win_rank : int, optional
-            Fixed rank threshold (required when *group_filter* is ``None``).
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters applied before aggregation.
-        group_filter : pl.Expr or list of pl.Expr, optional
-            Filter defining the comparison group.
-        status : {"Wins", "Losses"}, optional
-            Required when *group_filter* is provided.
-        top_k : int, optional
-            Limit the number of rows returned (group_filter mode only).
-
-        Returns
-        -------
-        pl.LazyFrame
-        """
-        if group_filter is None:
-            if win_rank is None:
-                raise ValueError("win_rank must be provided when group_filter is None.")
-            win_col = f"Win_at_rank{win_rank}"
-            group_level_win_losses = (
-                apply_filter(self.preaggregated_remaining_view, additional_filters)
-                .filter(pl.col(self.level) == "Arbitration")
-                .group_by(level)
-                .agg(Wins=pl.sum(win_col), Decisions=pl.sum("Decisions"))
-                .with_columns(Losses=pl.col("Decisions") - pl.col("Wins"))
-                .with_columns(
-                    Wins=pl.col("Wins") / pl.sum("Wins"),
-                    Losses=pl.col("Losses") / pl.sum("Losses"),
-                )
-            )
-
-            group_level_win_losses = group_level_win_losses.unpivot(
-                index=level,
-                on=["Wins", "Losses"],
-                variable_name="Status",
-                value_name="Percentage",
-            ).sort(["Status", "Percentage"], descending=True)
-
-            return group_level_win_losses
-
-        if status not in {"Wins", "Losses"}:
-            raise ValueError("When group_filter is provided, status must be either 'Wins' or 'Losses'.")
-
-        selected_group_rank_boundaries = self.get_selected_group_rank_boundaries(
-            group_filter=group_filter,
-            additional_filters=additional_filters,
-        )
-
-        stage_filtered_data = apply_filter(self.sample, additional_filters).filter(
-            pl.col(self.level).is_in(self.stages_from_arbitration_down)
-        )
-
-        if status == "Wins":
-            # Actions beaten by the selected group are ranked below the selected worst rank.
-            comparison_rows = stage_filtered_data.join(
-                selected_group_rank_boundaries,
-                on="Interaction ID",
-                how="inner",
-            ).filter(pl.col("Rank") > pl.col("selected_group_worst_rank"))
-        else:
-            # Actions that beat the selected group are ranked above the selected best rank.
-            comparison_rows = stage_filtered_data.join(
-                selected_group_rank_boundaries,
-                on="Interaction ID",
-                how="inner",
-            ).filter(pl.col("Rank") < pl.col("selected_group_best_rank"))
-
-        n_interactions = comparison_rows.select(pl.n_unique("Interaction ID")).collect().item()
-
-        distribution = (
-            comparison_rows.group_by(level)
-            .agg(Decisions=pl.len())
-            .filter(pl.col("Decisions") > 0)
-            .sort("Decisions", descending=True)
-        )
-        if top_k is not None:
-            distribution = distribution.head(top_k)
-
-        if n_interactions > 0:
-            distribution = distribution.with_columns(
-                (pl.col("Decisions") / pl.lit(n_interactions)).round(1).alias("Avg per Decision"),
-            )
-        else:
-            distribution = distribution.with_columns(
-                pl.lit(0.0).alias("Avg per Decision"),
-            )
-
-        return distribution
-
-    def _winning_from(
-        self,
-        interactions: pl.LazyFrame,
-        win_rank: int,
-        groupby_cols: list[str],
-        top_k: int = 20,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Actions beaten by the comparison group in its winning interactions.
-
-        Parameters
-        ----------
-        interactions
-            Interaction IDs where the comparison group wins.
-        win_rank
-            Rank threshold used to define "winning".
-        groupby_cols
-            Columns to group by (e.g. ``["Issue", "Group", "Action"]``).
-        top_k
-            Return only the top-k entries.
-        additional_filters
-            Optional extra filters (e.g. channel filter).
-        """
-        stage_filtered = apply_filter(self.sample, additional_filters).filter(
-            pl.col(self.level).is_in(self.stages_from_arbitration_down)
-        )
-        rows = stage_filtered.join(interactions, on="Interaction ID", how="inner").filter(pl.col("Rank") > win_rank)
-        n_interactions = rows.select(pl.n_unique("Interaction ID")).collect().item()
-        dist = (
-            rows.group_by(groupby_cols)
-            .agg(Decisions=pl.len())
-            .filter(pl.col("Decisions") > 0)
-            .sort("Decisions", descending=True)
-            .head(top_k)
-        )
-        if n_interactions > 0:
-            dist = dist.with_columns(
-                (pl.col("Decisions") / pl.lit(n_interactions)).round(1).alias("Avg per Decision"),
-            )
-        else:
-            dist = dist.with_columns(pl.lit(0.0).alias("Avg per Decision"))
-        return dist
-
-    def _losing_to(
-        self,
-        interactions: pl.LazyFrame,
-        win_rank: int,
-        groupby_cols: list[str],
-        top_k: int = 20,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Actions that beat the comparison group in its losing interactions.
-
-        Parameters
-        ----------
-        interactions
-            Interaction IDs where the comparison group loses.
-        win_rank
-            Rank threshold used to define "winning".
-        groupby_cols
-            Columns to group by (e.g. ``["Issue", "Group", "Action"]``).
-        top_k
-            Return only the top-k entries.
-        additional_filters
-            Optional extra filters (e.g. channel filter).
-        """
-        stage_filtered = apply_filter(self.sample, additional_filters).filter(
-            pl.col(self.level).is_in(self.stages_from_arbitration_down)
-        )
-        rows = stage_filtered.join(interactions, on="Interaction ID", how="inner").filter(pl.col("Rank") < win_rank)
-        n_interactions = rows.select(pl.n_unique("Interaction ID")).collect().item()
-        dist = (
-            rows.group_by(groupby_cols)
-            .agg(Decisions=pl.len())
-            .filter(pl.col("Decisions") > 0)
-            .sort("Decisions", descending=True)
-            .head(top_k)
-        )
-        if n_interactions > 0:
-            dist = dist.with_columns(
-                (pl.col("Decisions") / pl.lit(n_interactions)).round(1).alias("Avg per Decision"),
-            )
-        else:
-            dist = dist.with_columns(pl.lit(0.0).alias("Avg per Decision"))
-        return dist
-
-    def get_optionality_data(self, df: pl.LazyFrame | None = None, by_day: bool = False) -> pl.LazyFrame:
-        """Average number of actions per stage, optionally broken down by day.
-
-        Computes per-interaction action counts at each stage using
-        ``aggregate_remaining_per_stage``, then aggregates into a histogram.
-
-        Parameters
-        ----------
-        df : pl.LazyFrame, optional
-            Input data.  Defaults to :attr:`sample`.
-        by_day : bool, default False
-            If True, include ``"day"`` in the grouping for trend analysis.
-            When False, zero-action rows are injected for stages where some
-            interactions have no remaining actions.
-
-        Returns
-        -------
-        pl.LazyFrame
-        """
-        if df is None:
-            df = self.sample
-        expr = [
-            pl.len().alias("nOffers"),
-            pl.col("Propensity").filter(pl.col("Propensity") < 0.5).max().alias("bestPropensity"),
-        ]
-
-        group_by_columns = ["Interaction ID", "day"] if by_day else ["Interaction ID"]
-        outer_group = ["nOffers", self.level, "day"] if by_day else ["nOffers", self.level]
-
-        per_offer_count_and_stage = (
-            self.aggregate_remaining_per_stage(
-                df=df,
-                group_by_columns=group_by_columns,
-                aggregations=expr,
-            )
-            .group_by(outer_group)
-            .agg(Interactions=pl.len(), AverageBestPropensity=pl.mean("bestPropensity"))
-        )
-
-        if by_day:
-            return per_offer_count_and_stage.sort("nOffers", descending=True)
-
-        # Non-trend mode: inject zero-action rows for completeness
-        total_interactions = df.select("Interaction ID").collect().unique().height
-        schema = per_offer_count_and_stage.collect_schema()
-        zero_actions = (
-            per_offer_count_and_stage.group_by(self.level)
-            .agg(interaction_count=pl.sum("Interactions"))
-            .with_columns(
-                Interactions=(total_interactions - pl.col("interaction_count")).cast(schema["Interactions"]),
-                AverageBestPropensity=pl.lit(0.0).cast(schema["AverageBestPropensity"]),
-                nOffers=pl.lit(0).cast(schema["nOffers"]),
-            )
-            .drop("interaction_count")
-        )
-        optionality_data = pl.concat(
-            [
-                per_offer_count_and_stage,
-                zero_actions.select(per_offer_count_and_stage.collect_schema().names()),
-            ]
-        ).sort("nOffers", descending=True)
-
-        return optionality_data
-
-    def get_optionality_funnel(self, df: pl.LazyFrame | None = None) -> pl.LazyFrame:
-        """Optionality funnel: interaction counts bucketed by available-action count.
-
-        Buckets action counts into 0–6 and 7+, then counts interactions per
-        stage and bucket.  Used by the optionality funnel chart.
-
-        Parameters
-        ----------
-        df : pl.LazyFrame, optional
-            Input data.  Defaults to :attr:`sample`.
-
-        Returns
-        -------
-        pl.LazyFrame
-        """
-        if df is None:
-            df = self.sample
-        optionality_funnel = (
-            self.get_optionality_data(df)
-            .with_columns(
-                pl.when(pl.col("nOffers") >= 7)
-                .then(pl.lit("7+"))
-                .otherwise(pl.col("nOffers").cast(str))
-                .alias("available_actions")
-            )
-            .with_columns(
-                pl.col("available_actions").cast(pl.Enum(["0", "1", "2", "3", "4", "5", "6", "7+"])),
-            )
-            .group_by([self.level, "available_actions"])
-            .agg(pl.sum("Interactions"))
-            .sort([self.level, "available_actions"])
-        )
-        return optionality_funnel
-
-    def get_action_variation_data(
-        self,
-        stage: str,
-        color_by: str | None = None,
-    ) -> pl.LazyFrame:
-        """Get action variation data, optionally broken down by a categorical dimension.
-
-        Args:
-            stage: The stage to analyze
-            color_by: Optional categorical column to break down the variation by.
-                     Can use "Channel/Direction" to combine Channel and Direction columns.
-        """
-        # Handle combined Channel/Direction column
-        needs_channel_direction = color_by == "Channel/Direction"
-        if needs_channel_direction:
-            grouping_levels = ["Channel", "Direction", "Action"]
-        else:
-            grouping_levels = ["Action"] if color_by is None else [color_by, "Action"]
-
-        # Get distribution data grouped by Action (and optionally color_by dimension)
-        distribution = self.get_distribution_data(
-            stage=stage,
-            grouping_levels=grouping_levels,
-        ).collect()
-
-        # Create combined Channel/Direction column if needed
-        if needs_channel_direction:
-            distribution = distribution.with_columns(
-                pl.concat_str("Channel", "Direction", separator="/").alias("Channel/Direction")
-            )
-            color_by_col = "Channel/Direction"
-        else:
-            color_by_col = color_by
-
-        # Compute cumulative stats within each group if color_by is specified
-        if color_by is not None:
-            # Define final columns to keep
-            final_columns = [
-                "ActionIndex",
-                "Action",
-                "Decisions",
-                "cumDecisions",
-                "DecisionsFraction",
-                "ActionsFraction",
-                color_by_col,
-            ]
-
-            data = (
-                distribution.sort([color_by_col, "Decisions"], descending=[False, True])
-                .with_columns(cumDecisions=pl.col("Decisions").cum_sum().over(color_by_col).cast(pl.Int32))
-                .with_columns(
-                    DecisionsFraction=(pl.col("cumDecisions") / pl.col("Decisions").sum().over(color_by_col)),
-                    Decisions=pl.col("Decisions").cast(pl.Int32),
-                )
-                .with_columns(ActionIndex=pl.int_range(pl.len()).over(color_by_col) + 1)
-                .with_columns(
-                    ActionsFraction=pl.col("ActionIndex") / pl.col("ActionIndex").max().over(color_by_col),
-                    ActionIndex=pl.col("ActionIndex").cast(pl.UInt32),
-                    Decisions=pl.col("Decisions").cast(pl.UInt32),
-                    cumDecisions=pl.col("cumDecisions").cast(pl.UInt32),
-                )
-                .select(final_columns)
-            )
-
-            # Create zero points for each unique group value
-            unique_groups = data[color_by_col].unique()
-            zero_points = (
-                pl.DataFrame(
-                    {
-                        color_by_col: unique_groups,
-                        "ActionIndex": [0] * len(unique_groups),
-                        "Action": [""] * len(unique_groups),
-                        "Decisions": [0] * len(unique_groups),
-                        "cumDecisions": [0] * len(unique_groups),
-                        "DecisionsFraction": [0.0] * len(unique_groups),
-                        "ActionsFraction": [0.0] * len(unique_groups),
-                    }
-                )
-                .with_columns(
-                    pl.col("ActionIndex").cast(pl.UInt32),
-                    pl.col("Decisions").cast(pl.UInt32),
-                    pl.col("cumDecisions").cast(pl.UInt32),
-                )
-                .select(final_columns)
-            )
-
-            result = pl.concat([zero_points, data]).sort([color_by_col, "ActionIndex"])
-        else:
-            data = (
-                distribution.with_columns(cumDecisions=pl.col("Decisions").cum_sum().cast(pl.Int32))
-                .with_columns(
-                    DecisionsFraction=(pl.col("cumDecisions") / pl.sum("Decisions")),
-                    Decisions=pl.col("Decisions").cast(pl.Int32),
-                )
-                .with_row_index("ActionIndex", 1)
-                .with_columns(
-                    ActionsFraction=pl.col("ActionIndex") / pl.len(),
-                    ActionIndex=pl.col("ActionIndex").cast(pl.UInt32),
-                    Decisions=pl.col("Decisions").cast(pl.UInt32),
-                    cumDecisions=pl.col("cumDecisions").cast(pl.UInt32),
-                )
-            )
-
-            # Single zero point for non-grouped data
-            zero_point = pl.DataFrame(
-                {
-                    "ActionIndex": 0,
-                    "Action": "",
-                    "Decisions": 0,
-                    "cumDecisions": 0,
-                    "DecisionsFraction": 0.0,
-                    "ActionsFraction": 0.0,
-                }
-            ).with_columns(
-                pl.col("ActionIndex").cast(pl.UInt32),
-                pl.col("Decisions").cast(pl.UInt32),
-                pl.col("cumDecisions").cast(pl.UInt32),
-            )
-
-            result = pl.concat([zero_point, data])
-
-        return result.lazy()
-
-    def get_ab_test_results(self) -> pl.DataFrame:
-        """A/B test summary: control vs test counts and control percentage per stage.
-
-        Returns
-        -------
-        pl.DataFrame
-            One row per stage with columns for Control, Test counts and
-            Control Percentage. Rows preserve the canonical
-            ``AvailableNBADStages`` order (stages absent from the data are
-            omitted).
-        """
-        tbl = (
-            self.preaggregated_remaining_view.group_by(
-                # Tracked in docs/plans/decision-analyzer-TODO.md (P3: AB
-                # test: include IA properties) — IA properties not joined
-                # here when present.
-                [self.level, "Model Control Group"]
-            )
-            .agg(pl.len())
-            .collect()
-            .pivot(
-                index=self.level,
-                values="len",
-                columns=["Model Control Group"],
-                sort_columns=True,
-            )
-            .with_columns((pl.col("Control") / (pl.col("Test") + pl.col("Control"))).alias("Control Percentage"))
-        )
-        stage_order = pl.DataFrame(
-            {self.level: self.AvailableNBADStages, "_stage_order": list(range(len(self.AvailableNBADStages)))}
-        )
-        tbl = (
-            tbl.join(stage_order, on=self.level, how="left").sort("_stage_order", nulls_last=True).drop("_stage_order")
-        )
-        return tbl
-
-    def get_thresholding_data(self, fld: str, quantile_range=range(10, 100, 10)) -> pl.DataFrame:
-        """Quantile-based thresholding analysis at Arbitration.
-
-        Computes counts and threshold values at each quantile for the given
-        field (*fld*).  Results are cached per ``(fld, quantile_range)``.
-
-        Parameters
-        ----------
-        fld : str
-            Column name to compute quantiles for (e.g. ``"Propensity"``).
-        quantile_range : range, default ``range(10, 100, 10)``
-            Percentile breakpoints to compute.
-
-        Returns
-        -------
-        pl.DataFrame
-            Long-format table with columns ``Decile``, ``Count``,
-            ``Threshold``, and the stage-level column.
-        """
-        cache_key = (fld, tuple(quantile_range))
-        if cache_key in self._thresholding_cache:
-            return self._thresholding_cache[cache_key]
-
-        thresholds_wide = (
-            # Note: runs on pre-aggregated data - maybe should be using _min/_max for filtering instead
-            self.preaggregated_filter_view.filter(pl.col(self.level).is_in(self.stages_from_arbitration_down))
-            .select(
-                # TODO can probably code this up more efficiently
-                [pl.col(fld).explode().quantile(q / 100.0).alias(f"p{q}") for q in quantile_range]
-                + [
-                    ((pl.col(fld).explode()) < (pl.col(fld).explode().quantile(q / 100.0))).sum().alias(f"n{q}")
-                    for q in quantile_range
-                ]
-                + [pl.lit("Arbitration").alias(self.level)]
-            )
-            .collect()
-        )
-        thresholds_long = (
-            thresholds_wide.select([self.level] + [f"n{q}" for q in quantile_range])
-            .unpivot(
-                index=self.level,
-                on=cs.numeric(),
-                variable_name="Decile",
-                value_name="Count",
-            )
-            .with_columns(pl.col("Decile").str.replace("n", "p"))
-            .join(
-                thresholds_wide.select([self.level] + [f"p{q}" for q in quantile_range]).unpivot(
-                    index=self.level,
-                    on=cs.numeric(),
-                    variable_name="Decile",
-                    value_name="Threshold",
-                ),
-                on=[self.level, "Decile"],
-            )
-            .sort([self.level, "Threshold"])
-        ).filter(pl.col(self.level) == "Arbitration")
-        self._thresholding_cache[cache_key] = thresholds_long
-        return thresholds_long
-
-    def priority_component_distribution(
-        self,
-        component: str,
-        granularity: str,
-        stage: str | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Data for a single component's distribution, grouped by granularity.
-
-        Parameters
-        ----------
-        component : str
-            Column name of the component to analyze.
-        granularity : str
-            Column to group by (e.g. "Issue", "Group", "Action").
-        stage : str, optional
-            Filter to actions remaining at this stage. If None, uses all
-            rows with non-null Priority.
-        additional_filters : pl.Expr or list[pl.Expr], optional
-            Extra filters applied to the sample (e.g. channel filter).
-        """
-        cols = [granularity, component]
-        df = self._remaining_at_stage(stage, additional_filters=additional_filters)
-        return df.select(cols).sort(granularity)
-
-    def all_components_distribution(
-        self,
-        granularity: str,
-        stage: str | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Data for the overview panel: all prioritization components at once.
-
-        Parameters
-        ----------
-        granularity : str
-            Column to group by.
-        stage : str, optional
-            Filter to actions remaining at this stage.
-        additional_filters : pl.Expr or list[pl.Expr], optional
-            Extra filters applied to the sample (e.g. channel filter).
-        """
-        from .utils import PRIO_COMPONENTS
-
-        available = set(self.sample.collect_schema().names())
-        cols = [c for c in PRIO_COMPONENTS if c in available]
-        df = self._remaining_at_stage(stage, additional_filters=additional_filters)
-        return df.select([granularity] + cols).sort(granularity)
-
-    def _remaining_at_stage(
-        self,
-        stage: str | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Return sample rows remaining at *stage*.
-
-        Uses the ``aggregate_remaining_per_stage`` logic: rows whose stage
-        order is >= the selected stage are "remaining" there.  If *stage*
-        is None, falls back to rows with non-null Priority.
-        """
-        base = apply_filter(self.sample, additional_filters)
-        if stage is None:
-            return base.filter(pl.col("Priority").is_not_null())
-        stage_idx = self.AvailableNBADStages.index(stage) if stage in self.AvailableNBADStages else 0
-        remaining_stages = self.AvailableNBADStages[stage_idx:]
-        return base.filter(pl.col(self.level).is_in(remaining_stages))
-
-    def aggregate_remaining_per_stage(
-        self,
-        df: pl.LazyFrame,
-        group_by_columns: list[str],
-        aggregations: list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """
-        Workhorse function to convert the raw Decision Analyzer data (filter view) to
-        the aggregates remaining per stage, ensuring all stages are represented.
-        """
-        if aggregations is None:
-            aggregations = []
-        stage_orders = (
-            df.group_by(self.level)
-            .agg(pl.min("Stage Order"))
-            .with_columns(
-                pl.col(self.level)
-                .cast(pl.Utf8)
-                .cast(
-                    pl.Enum(self.AvailableNBADStages)
-                )  # weird polars behaviour(version: 1.29), try removing in later patches
-            )
-        )
-
-        def aggregate_over_remaining_stages(df, stage, remaining_stages):
-            return (
-                df.filter(pl.col(self.level).is_in(remaining_stages))
-                .group_by(group_by_columns)
-                .agg(aggregations)
-                .with_columns(pl.lit(stage).alias(self.level))
-            )
-
-        aggs = {
-            stage: aggregate_over_remaining_stages(df, stage, self.AvailableNBADStages[i:])
-            for (i, stage) in enumerate(self.AvailableNBADStages)
-        }
-        remaining_view = (
-            pl.concat(aggs.values())
-            .with_columns(pl.col(self.level).cast(stage_orders.collect_schema()[self.level]))
-            .join(stage_orders, on=self.level, how="left")
-        )
-
-        return remaining_view
-
-    def filtered_action_counts(
-        self,
-        groupby_cols: list[str],
-        propensity_th: float | None = None,
-        priority_th: float | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Return action counts from the sample, optionally classified by propensity/priority thresholds.
-
-        Parameters
-        ----------
-        groupby_cols : list of str
-            Column names to group by.
-        propensity_th : float, optional
-            Propensity threshold for classifying offers.
-        priority_th : float, optional
-            Priority threshold for classifying offers.
-        additional_filters : pl.Expr or list[pl.Expr], optional
-            Extra filters applied to the sample (e.g. channel filter).
-
-        Returns
-        -------
-        pl.LazyFrame
-            Aggregated action counts per group, with quality buckets when
-            both thresholds are provided.
-        """
-        df = apply_filter(self.sample, additional_filters)
-        additional_cols = ["Action", "Propensity"]
-        required_cols = list(set(groupby_cols + additional_cols))
-        for col in required_cols:
-            if col not in df.collect_schema().names():
-                raise ValueError(f"Column '{col}' not found in the dataframe.")
-
-        if propensity_th is None or priority_th is None:
-            return df.group_by(groupby_cols).agg(
-                no_of_offers=pl.count("Action"),
-            )
-
-        # Only classify actions by propensity/priority at stages where propensity is available
-        # Actions filtered before ActionPropensity stage don't have meaningful propensity scores
-        has_propensity = pl.col(self.level).is_in(self.stages_with_propensity)
-
-        propensity_classifying_expr = [
-            pl.col("Action")
-            .filter(has_propensity & (pl.col("Propensity") == 0.5) & (pl.col(self.level) != "Output"))
-            .count()
-            .alias("new_models"),
-            pl.col("Action")
-            .filter(has_propensity & (pl.col("Propensity") < propensity_th) & (pl.col("Propensity") != 0.5))
-            .count()
-            .alias("poor_propensity_offers"),
-            pl.col("Action")
-            .filter(has_propensity & (pl.col("Priority") < priority_th) & (pl.col("Propensity") != 0.5))
-            .count()
-            .alias("poor_priority_offers"),
-            pl.col("Action")
-            .filter(
-                has_propensity
-                & (pl.col("Propensity") >= propensity_th)
-                & (pl.col("Propensity") != 0.5)
-                & (pl.col("Priority") >= priority_th)
-            )
-            .count()
-            .alias("good_offers"),
-        ]
-        return df.group_by(groupby_cols).agg(no_of_offers=pl.count("Action"), *propensity_classifying_expr)
-
-    def get_offer_quality(self, df: pl.LazyFrame, group_by: str | list[str]) -> pl.LazyFrame:
-        """Cumulative offer-quality breakdown across stages.
-
-        Takes a filtered-action-counts frame (from :meth:`filtered_action_counts`)
-        and converts it to a remaining-per-stage view, joining in customers
-        that have zero actions so they are counted as well.
-
-        Parameters
-        ----------
-        df : pl.LazyFrame
-            Filtered action counts with columns ``no_of_offers``,
-            ``new_models``, ``poor_propensity_offers``, etc.
-        group_by : str or list of str
-            Columns to group by (e.g. ``["Interaction ID"]``).
-
-        Returns
-        -------
-        pl.LazyFrame
-            Per-stage quality classification with boolean flag columns
-            (``has_no_offers``, ``atleast_one_relevant_action``, etc.).
-        """
-        # Get all unique customers/interactions from the sample to ensure we count those with no actions
-        if isinstance(group_by, str):
-            group_by = [group_by]
-
-        # Get the full set of customers from the FIRST stage (where all customers start)
-        # This ensures we track all customers throughout the pipeline
-        first_stage = self.AvailableNBADStages[0]
-        all_customers = self.sample.filter(pl.col(self.level) == first_stage).select(group_by).unique().collect().lazy()
-
-        dfs = []
-        for i, stage in enumerate(self.AvailableNBADStages):
-            # Get aggregated counts for this stage (only interactions WITH actions)
-            stage_df = (
-                # Tracked in docs/plans/decision-analyzer-TODO.md (P2:
-                # Refactor get_offer_quality) — delegate to
-                # aggregate_remaining_per_stage instead of this manual loop.
-                df.filter(pl.col(self.level).is_in(self.AvailableNBADStages[i:]))
-                .group_by(group_by)
-                .agg(
-                    pl.sum(
-                        "no_of_offers",
-                        "new_models",
-                        "poor_propensity_offers",
-                        "poor_priority_offers",
-                        "good_offers",
-                    )
-                )
-                .with_columns(pl.lit(stage).alias(self.level))
-            )
-
-            # Left join with all customers to include those with no actions (fill with 0s)
-            # Cast stage_df level column to categorical to match join keys
-            stage_df = stage_df.with_columns(pl.col(self.level).cast(pl.Categorical))
-
-            # Add stage to all customers and join
-            stage_customers = all_customers.with_columns(pl.lit(stage).cast(pl.Categorical).alias(self.level))
-            stage_df = stage_customers.join(stage_df, on=group_by + [self.level], how="left").with_columns(
-                pl.col(
-                    "no_of_offers", "new_models", "poor_propensity_offers", "poor_priority_offers", "good_offers"
-                ).fill_null(0)
-            )
-
-            dfs.append(stage_df)
-        stage_df = pl.concat(dfs)
-
-        # Determine if stage has propensity scores available (ActionPropensity and later)
-        has_propensity = pl.col(self.level).is_in(self.stages_with_propensity)
-
-        stage_df = stage_df.with_columns(
-            has_no_offers=pl.when(pl.col("no_of_offers") == 0).then(pl.lit(1)).otherwise(pl.lit(0)),
-            # For stages WITH propensity: classify as relevant (green) or irrelevant (orange)
-            atleast_one_relevant_action=pl.when(has_propensity & (pl.col("good_offers") >= 1))
-            .then(pl.lit(1))
-            .otherwise(pl.lit(0)),
-            only_irrelevant_actions=pl.when(
-                has_propensity & (pl.col("good_offers") == 0) & (pl.col("no_of_offers") > 0)
-            )
-            .then(pl.lit(1))
-            .otherwise(0),
-            # For stages WITHOUT propensity: just show as having actions (blue)
-            atleast_one_action=pl.when(~has_propensity & (pl.col("no_of_offers") > 0))
-            .then(pl.lit(1))
-            .otherwise(pl.lit(0)),
-        )
-        return stage_df
 
     @cached_property
     def overview_stats(self) -> dict[str, object]:
@@ -2343,7 +1014,7 @@ class DecisionAnalyzer:
         """
 
         nOffersPerStage = (
-            self.get_optionality_data(self.sample)
+            self.aggregates.get_optionality_data(self.sample)
             .group_by(self.level)
             .agg(pl.col("nOffers").mean().round().cast(pl.Int16))
             .collect()
@@ -2397,545 +1068,3 @@ class DecisionAnalyzer:
         )
 
         return {k: kpis[k].item() for k in kpis.columns}
-
-    def get_sensitivity(
-        self,
-        win_rank: int = 1,
-        group_filter: pl.Expr | list[pl.Expr] | None = None,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Global or local sensitivity of the prioritization factors.
-
-        Parameters
-        ----------
-        win_rank : int
-            Maximum rank to be considered a winner.
-        group_filter : pl.Expr, optional
-            Selected offers, only used in local sensitivity analysis.
-            When ``None`` (global), results are cached by ``win_rank``.
-        additional_filters : pl.Expr or list[pl.Expr], optional
-            Extra filters applied to the sample before re-ranking
-            (e.g. channel filter).  When set, caching is bypassed.
-
-        Returns
-        -------
-        pl.LazyFrame
-        """
-        is_global_sensitivity = group_filter is None
-        if is_global_sensitivity and additional_filters is None:
-            if win_rank in self._sensitivity_cache:
-                return self._sensitivity_cache[win_rank]
-            group_filter = pl.col("Rank") <= win_rank
-        elif is_global_sensitivity:
-            group_filter = pl.col("Rank") <= win_rank
-
-        sensitivity = (
-            apply_filter(
-                self.re_rank(additional_filters=additional_filters), group_filter
-            )  # don't put filters in rerank function, we need to filter after reranking!
-            # .filter(pl.col("rank_PVCL") <= win_rank)
-            .select(
-                [
-                    pl.col("Interaction ID")
-                    .filter(pl.col(x) <= win_rank)
-                    .n_unique()
-                    .cast(pl.Int32)  # thinks they are unsigned int, 33-34 returns big number
-                    .alias((f"{x.split('_')[1]}_win_count"))  # calculating win_counts of different combinations
-                    for x in [
-                        "rank_PVCL",
-                        "rank_VCL",
-                        "rank_PCL",
-                        "rank_PVL",
-                        "rank_PVC",
-                    ]
-                ]
-            )
-            .with_columns(
-                [
-                    (pl.col("PVCL_win_count") - pl.col(x)).alias(x)
-                    for x in [
-                        "VCL_win_count",
-                        "PCL_win_count",
-                        "PVL_win_count",
-                        "PVC_win_count",
-                    ]
-                ]
-            )
-            .rename(
-                {
-                    "PVCL_win_count": "Priority",
-                    "VCL_win_count": "Propensity",
-                    "PCL_win_count": "Value",
-                    "PVL_win_count": "Context Weights",
-                    "PVC_win_count": "Levers",
-                }
-            )
-            .unpivot(variable_name="Factor", value_name="Influence")
-        )
-        if is_global_sensitivity:
-            sensitivity = sensitivity.with_columns(Influence=pl.col("Influence").abs())
-            if additional_filters is None:
-                self._sensitivity_cache[win_rank] = sensitivity
-        return sensitivity
-
-    def get_offer_variability_stats(self, stage: str) -> dict[str, float]:
-        """Summary statistics for action variation at a stage.
-
-        Parameters
-        ----------
-        stage : str
-            Stage to analyse.
-
-        Returns
-        -------
-        dict
-            ``n90`` — number of actions covering 90 % of decisions.
-            ``gini`` — Gini coefficient of decision concentration.
-        """
-        offer_variability_data = self.get_action_variation_data(stage)
-        return {
-            "n90": bisect_left(
-                offer_variability_data.select("DecisionsFraction").collect()["DecisionsFraction"],
-                0.9,
-            )
-            - 1,  # first one is dummy
-            "gini": 1.0
-            - gini_coefficient(
-                offer_variability_data.select(["ActionsFraction", "DecisionsFraction"]).collect(),
-                "ActionsFraction",
-                "DecisionsFraction",
-            ),
-        }
-
-    def get_winning_or_losing_interactions(
-        self,
-        group_filter: pl.Expr | list[pl.Expr],
-        win: bool,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.LazyFrame:
-        """Interaction IDs where the comparison group wins or loses.
-
-        Parameters
-        ----------
-        group_filter : pl.Expr or list of pl.Expr
-            Filter defining the comparison group.
-        win : bool
-            If True, return interactions where the group wins (there are
-            lower-ranked actions outside the group).  If False, return
-            interactions where the group loses (there are higher-ranked
-            actions outside the group).
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters (e.g. channel filter).
-
-        Returns
-        -------
-        pl.LazyFrame
-            Single-column frame of unique ``Interaction ID`` values.
-        """
-        selected_group_rank_boundaries = self.get_selected_group_rank_boundaries(
-            group_filter=group_filter,
-            additional_filters=additional_filters,
-        )
-        stage_filtered_data = apply_filter(self.sample, additional_filters).filter(
-            pl.col(self.level).is_in(self.stages_from_arbitration_down)
-        )
-
-        if win:
-            interaction_filter = pl.col("Rank") > pl.col("selected_group_worst_rank")
-        else:
-            interaction_filter = pl.col("Rank") < pl.col("selected_group_best_rank")
-
-        return (
-            stage_filtered_data.join(selected_group_rank_boundaries, on="Interaction ID", how="inner")
-            .filter(interaction_filter)
-            .select(pl.col("Interaction ID").unique())
-        )
-
-    def get_win_loss_counts(
-        self,
-        group_filter: pl.Expr | list[pl.Expr],
-        win_rank: int = 1,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> dict[str, int]:
-        """Count wins and losses for the comparison group at a given rank threshold.
-
-        A **win** is an interaction where at least one member of the comparison
-        group achieves a rank of *win_rank* or better (lower). A **loss** is
-        any interaction where the group participates but none rank that high.
-
-        Parameters
-        ----------
-        group_filter
-            Filter expression(s) defining the comparison group.
-        win_rank
-            Rank threshold. The group "wins" when its best rank <= win_rank.
-        additional_filters
-            Optional extra filters (e.g. channel/direction).
-
-        Returns
-        -------
-        dict with keys ``"wins"``, ``"losses"``, ``"total"``.
-        """
-        boundaries = self.get_selected_group_rank_boundaries(
-            group_filter=group_filter,
-            additional_filters=additional_filters,
-        )
-        counts = boundaries.select(
-            wins=(pl.col("selected_group_best_rank") <= win_rank).sum(),
-            losses=(pl.col("selected_group_best_rank") > win_rank).sum(),
-            total=pl.len(),
-        ).collect()
-
-        row = counts.row(0, named=True)
-        return {"wins": row["wins"], "losses": row["losses"], "total": row["total"]}
-
-    def get_win_loss_distributions(
-        self,
-        interactions_win: pl.LazyFrame,
-        interactions_loss: pl.LazyFrame,
-        groupby_cols: list[str],
-        top_k: int,
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-        group_filter: pl.Expr | list[pl.Expr] | None = None,
-        win_rank: int | None = None,
-    ) -> tuple[pl.LazyFrame, pl.LazyFrame]:
-        """Distribution of actions the comparison group wins from and loses to.
-
-        Computes two aggregated distributions in a single pass over the
-        stage-filtered data:
-
-        * **winning_from**: actions ranked below the comparison group (i.e.
-          actions it beats).
-        * **losing_to**: actions ranked above the comparison group (i.e.
-          actions that beat it).
-
-        Either *group_filter* or *win_rank* must be provided to define the
-        rank boundary.  When *group_filter* is given the boundary is the
-        per-interaction best/worst rank of the selected group.  When only
-        *win_rank* is given a fixed rank threshold is used instead.
-
-        Parameters
-        ----------
-        interactions_win
-            Interaction IDs where the comparison group wins (from
-            ``get_winning_or_losing_interactions(win=True)``).
-        interactions_loss
-            Interaction IDs where the comparison group loses (from
-            ``get_winning_or_losing_interactions(win=False)``).
-        groupby_cols
-            Columns to group by (e.g. ``["Action"]``).
-        top_k
-            Return only the top-k entries per distribution.
-        additional_filters
-            Optional extra filters (e.g. channel/direction).
-        group_filter
-            Filter expression(s) defining the comparison group.
-        win_rank
-            Fixed rank threshold. Required when *group_filter* is ``None``.
-
-        Returns
-        -------
-        tuple of (winning_from, losing_to), both ``pl.LazyFrame`` with
-        columns from *groupby_cols* plus ``Decisions``.
-        """
-        stage_filtered_data = apply_filter(self.sample, additional_filters).filter(
-            pl.col(self.level).is_in(self.stages_from_arbitration_down)
-        )
-
-        if group_filter is None:
-            if win_rank is None:
-                raise ValueError("win_rank must be provided when group_filter is None.")
-            win_rows = stage_filtered_data.filter(pl.col("Rank") > win_rank)
-            loss_rows = stage_filtered_data.filter(pl.col("Rank") < win_rank)
-        else:
-            boundaries = self.get_selected_group_rank_boundaries(
-                group_filter=group_filter,
-                additional_filters=additional_filters,
-            )
-            joined = stage_filtered_data.join(boundaries, on="Interaction ID", how="inner")
-            win_rows = joined.filter(pl.col("Rank") > pl.col("selected_group_worst_rank"))
-            loss_rows = joined.filter(pl.col("Rank") < pl.col("selected_group_best_rank"))
-
-        def _aggregate(rows: pl.LazyFrame, interactions: pl.LazyFrame) -> pl.LazyFrame:
-            return (
-                rows.join(interactions, on="Interaction ID", how="inner")
-                .group_by(groupby_cols)
-                .agg(Decisions=pl.len())
-                .sort("Decisions", descending=True)
-                .filter(pl.col("Decisions") > 0)
-                .head(top_k)
-            )
-
-        return _aggregate(win_rows, interactions_win), _aggregate(loss_rows, interactions_loss)
-
-    def get_win_distribution_data(
-        self,
-        lever_condition: pl.Expr,
-        lever_value: float | None = None,
-        all_interactions: int | None = None,
-    ) -> pl.DataFrame:
-        """
-        Calculate win distribution data for business lever analysis.
-
-        This method generates distribution data showing how actions perform in
-        arbitration decisions, both in baseline conditions and optionally with
-        lever adjustments applied.
-
-        Parameters
-        ----------
-        lever_condition : pl.Expr
-            Polars expression defining which actions to apply the lever to.
-            Example: pl.col("Action") == "SpecificAction" or
-                    (pl.col("Issue") == "Service") & (pl.col("Group") == "Cards")
-        lever_value : float, optional
-            The lever multiplier value to apply to selected actions.
-            If None, returns baseline distribution only.
-            If provided, returns both original and lever-adjusted win counts.
-        all_interactions : int, optional
-            Total number of interactions to calculate "no winner" count.
-            If provided, enables calculation of interactions without any winner.
-            If None, "no winner" data is not calculated.
-
-        Returns
-        -------
-        pl.DataFrame
-            DataFrame containing win distribution with columns:
-            - Issue, Group, Action: Action identifiers
-            - original_win_count: Number of rank-1 wins in baseline scenario
-            - new_win_count: Number of rank-1 wins after lever adjustment (only if lever_value provided)
-            - n_decisions_survived_to_arbitration: Number of arbitration decisions the action participated in
-            - selected_action: "Selected" for actions matching lever_condition, "Rest" for others
-            - no_winner_count: Number of interactions without any winner (only if all_interactions provided)
-
-        Notes
-        -----
-        - Only includes actions that survive to arbitration stage
-        - Win counts represent rank-1 (first place) finishes in arbitration decisions
-        - This is a zero-sum analysis: boosting selected actions suppresses others
-        - Results are sorted by win count (new_win_count if available, else original_win_count)
-        - When all_interactions is provided, "no winner" represents interactions without any rank-1 winner
-
-        Examples
-        --------
-        Get baseline distribution for a specific action:
-        >>> lever_cond = pl.col("Action") == "MyAction"
-        >>> baseline = decision_analyzer.get_win_distribution_data(lever_cond)
-
-        Get distribution with 2x lever applied to service actions:
-        >>> lever_cond = pl.col("Issue") == "Service"
-        >>> with_lever = decision_analyzer.get_win_distribution_data(lever_cond, 2.0)
-
-        Get distribution with no winner count:
-        >>> total_interactions = 10000
-        >>> with_no_winner = decision_analyzer.get_win_distribution_data(lever_cond, 2.0, total_interactions)
-        """
-        if lever_value is None:
-            # Return baseline distribution only
-            original_winners = self.re_rank(
-                additional_filters=pl.col(self.level).is_in(self.stages_from_arbitration_down),
-            ).select(["Issue", "Group", "Action"] + ["Interaction ID", "Rank"])
-
-            result = (
-                original_winners.group_by(["Issue", "Group", "Action"])
-                .agg(
-                    original_win_count=pl.col("Rank").filter(pl.col("Rank") == 1).len(),
-                    n_decisions_survived_to_arbitration=pl.col("Interaction ID").n_unique(),
-                )
-                .with_columns(
-                    selected_action=pl.when(lever_condition).then(pl.lit("Selected")).otherwise(pl.lit("Rest"))
-                )
-                .collect()
-                .sort("original_win_count", descending=True)
-            )
-
-            # Add no winner count if all_interactions is provided
-            if all_interactions is not None:
-                winner_df = original_winners.select("Interaction ID").collect()
-                interactions_with_winners = winner_df.n_unique()
-                no_winner_count = max(0, all_interactions - interactions_with_winners)  # Ensure non-negative
-
-                # Create a row with the same data types as the result
-                no_winner_data = {
-                    "Issue": ["No Winner"],
-                    "Group": ["No Winner"],
-                    "Action": ["No Winner"],
-                    "original_win_count": [no_winner_count],
-                    "n_decisions_survived_to_arbitration": [0],
-                    "selected_action": ["No Winner"],
-                }
-
-                # Cast to match result schema
-                no_winner_row = pl.DataFrame(no_winner_data).cast({k: v for k, v in result.schema.items()})
-
-                result = pl.concat([result, no_winner_row])
-
-            return result
-        else:
-            # Return both baseline and lever-adjusted distribution
-            recalculated_winners = self.re_rank(
-                overrides=[
-                    (pl.when(lever_condition).then(pl.lit(lever_value)).otherwise(pl.col("Levers"))).alias("Levers")
-                ],
-                additional_filters=pl.col(self.level).is_in(self.stages_from_arbitration_down),
-            ).select(["Issue", "Group", "Action"] + ["Interaction ID", "Rank", "rank_PVCL"])
-
-            result_lf = (
-                recalculated_winners.group_by(["Issue", "Group", "Action"])
-                .agg(
-                    original_win_count=pl.col("Rank").filter(pl.col("Rank") == 1).len(),
-                    new_win_count=pl.col("rank_PVCL").filter(pl.col("rank_PVCL") == 1).len(),
-                    n_decisions_survived_to_arbitration=pl.col("Interaction ID").n_unique(),
-                )
-                .with_columns(
-                    selected_action=pl.when(lever_condition).then(pl.lit("Selected")).otherwise(pl.lit("Rest"))
-                )
-            )
-            result = result_lf.collect().sort("new_win_count", descending=True)
-
-            # Add no winner count if all_interactions is provided
-            if all_interactions is not None:
-                # Calculate no winner count based on new ranking
-                new_winner_df = recalculated_winners.filter(pl.col("rank_PVCL") == 1).select("Interaction ID").collect()
-                interactions_with_new_winners = new_winner_df.n_unique()
-                no_winner_count = max(0, all_interactions - interactions_with_new_winners)  # Ensure non-negative
-
-                # Create a row with the same data types as the result
-                no_winner_data = {
-                    "Issue": ["No Winner"],
-                    "Group": ["No Winner"],
-                    "Action": ["No Winner"],
-                    "original_win_count": [0],  # No winner has no original wins
-                    "new_win_count": [no_winner_count],
-                    "n_decisions_survived_to_arbitration": [0],
-                    "selected_action": ["No Winner"],
-                }
-
-                # Cast to match result schema
-                no_winner_row = pl.DataFrame(no_winner_data).cast({k: v for k, v in result.schema.items()})
-
-                result = pl.concat([result, no_winner_row])
-
-            return result
-
-    def get_trend_data(
-        self,
-        stage: str = "AvailableActions",
-        scope: Literal["Group", "Issue", "Action"] | None = "Group",
-        additional_filters: pl.Expr | list[pl.Expr] | None = None,
-    ) -> pl.DataFrame:
-        """Daily trend of unique decisions from a given stage onward.
-
-        Parameters
-        ----------
-        stage : str, default "AvailableActions"
-            Starting stage; all stages from this point onward are included.
-        scope : {"Group", "Issue", "Action"} or None, default "Group"
-            Optional grouping dimension.  If ``None``, returns totals by day.
-        additional_filters : pl.Expr or list of pl.Expr, optional
-            Extra filters applied to the sample.
-
-        Returns
-        -------
-        pl.DataFrame
-            Columns: ``day``, optionally *scope*, and ``Decisions``.
-        """
-        stages = self.AvailableNBADStages[self.AvailableNBADStages.index(stage) :]
-        group_by = ["day"] if scope is None else ["day", scope]
-
-        trend_data = (
-            apply_filter(self.sample, additional_filters)
-            .filter(pl.col(self.level).is_in(stages))
-            .group_by(group_by)
-            .agg(pl.n_unique("Interaction ID").alias("Decisions"))
-            .sort(group_by)
-        )
-
-        return trend_data
-
-    def find_lever_value(
-        self,
-        lever_condition: pl.Expr,
-        target_win_percentage: float,
-        win_rank: int = 1,
-        low: float = 0,
-        high: float = 100,
-        precision: float = 0.01,
-        ranking_stages: list[str] | None = None,
-    ) -> float:
-        """
-        Binary search algorithm to find lever value needed to achieve a desired win percentage.
-
-        Parameters
-        ----------
-        lever_condition : pl.Expr
-            Polars expression that defines which actions should receive the lever
-        target_win_percentage : float
-            The desired win percentage (0-100)
-        win_rank : int, default 1
-            Consider actions winning if they rank <= this value
-        low : float, default 0
-            Lower bound for lever search range
-        high : float, default 100
-            Upper bound for lever search range
-        precision : float, default 0.01
-            Search precision - smaller values give more accurate results
-        ranking_stages : list[str], optional
-            List of stages to include in analysis. Defaults to ["Arbitration"]
-
-        Returns
-        -------
-        float
-            The lever value needed to achieve the target win percentage
-
-        Raises
-        ------
-        ValueError
-            If the target win percentage cannot be achieved within the search range
-        """
-        if ranking_stages is None:
-            ranking_stages = ["Arbitration"]
-
-        def _calculate_action_win_percentage(lever: float) -> float:
-            """Calculate win percentage for a given lever value"""
-            ranked_df = self.re_rank(
-                overrides=[(pl.when(lever_condition).then(pl.lit(lever)).otherwise(pl.col("Levers"))).alias("Levers")],
-                additional_filters=pl.col(self.level).is_in(self.stages_from_arbitration_down),
-            ).filter(pl.col("rank_PVCL") <= win_rank)
-
-            selected_wins_df = ranked_df.filter(lever_condition).select("Interaction ID").collect()
-            selected_wins = selected_wins_df.height
-            selected_total_df = ranked_df.select("Interaction ID").collect()
-            selected_total = selected_total_df.height
-            percentage = (selected_wins / selected_total) * 100
-            return percentage
-
-        beginning_high = high
-        beginning_low = low
-
-        # Check if target is achievable within bounds
-        low_percentage = _calculate_action_win_percentage(beginning_low)
-        high_percentage = _calculate_action_win_percentage(beginning_high)
-
-        if target_win_percentage < low_percentage:
-            raise ValueError(
-                f"Target {target_win_percentage}% is too low. Even at lever {beginning_low}, your actions win in {low_percentage:.1f}% of interactions at arbitration."
-                "You might have interactions where only your selected actions survive until arbitration. So they will win no matter what."
-            )
-        elif target_win_percentage > high_percentage:
-            raise ValueError(
-                f"Target {target_win_percentage}% is too high. Even at lever {beginning_high}, you only get {high_percentage:.1f}%. "
-                f"You can increase the search range."
-            )
-
-        while high - low > precision:
-            mid = (low + high) / 2
-
-            current_win_percentage = _calculate_action_win_percentage(mid)
-
-            if current_win_percentage < target_win_percentage:
-                low = mid
-            else:
-                high = mid
-
-        final_lever = (low + high) / 2
-        return final_lever
